fix: move nav folder from integration/nav to integration/pkg/api
