+++ integration/oasis/stationfinder/basic_finder_test.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 	"testing"
 
-	"github.com/Telenav/osrm-backend/integration/nav"
+	"github.com/Telenav/osrm-backend/integration/pkg/api/nav"
 	"github.com/Telenav/osrm-backend/integration/pkg/api/search/nearbychargestation"
 )
