+++ integration/oasis/stationfinder/low_energy_location_station_finder.go
@@ -1,9 +1,9 @@
 package stationfinder
 
 import (
-	"github.com/Telenav/osrm-backend/integration/nav"
 	"github.com/Telenav/osrm-backend/integration/oasis/searchconnector"
 	"github.com/Telenav/osrm-backend/integration/oasis/searchhelper"
+	"github.com/Telenav/osrm-backend/integration/pkg/api/nav"
 	"github.com/Telenav/osrm-backend/integration/pkg/api/search/searchcoordinate"
 )
